[tb/fix-label-image-popover] Move relevant files to label image editor folder
