+++ packages/perseus-editor/src/widgets/label-image-editor.tsx
@@ -5,12 +5,12 @@ import * as React from "react";
 
 import FormWrappedTextField from "../components/form-wrapped-text-field";
 
-import AnswerChoices from "./label-image/answer-choices";
-import Behavior from "./label-image/behavior";
-import QuestionMarkers from "./label-image/question-markers";
-import SelectImage from "./label-image/select-image";
+import AnswerChoices from "./label-image-editor/answer-choices";
+import Behavior from "./label-image-editor/behavior";
+import QuestionMarkers from "./label-image-editor/question-markers";
+import SelectImage from "./label-image-editor/select-image";
 
-import type {PreferredPopoverDirection} from "./label-image/behavior";
+import type {PreferredPopoverDirection} from "./label-image-editor/behavior";
 import type {
     PerseusLabelImageWidgetOptions,
     LabelImageDefaultWidgetOptions,
