+++ packages/perseus-editor/src/widgets/label-image-editor/__docs__/answer-choices.stories.tsx
