+++ packages/perseus-editor/src/widgets/label-image-editor/__docs__/behavior.stories.tsx
