+++ packages/perseus-editor/src/widgets/label-image-editor/__docs__/marker.stories.tsx
