+++ packages/perseus-editor/src/widgets/label-image-editor/__docs__/question-markers.stories.tsx
