+++ packages/perseus-editor/src/widgets/label-image-editor/__docs__/select-image.stories.tsx
