+++ packages/perseus-editor/src/widgets/label-image-editor/answer-choices.tsx
