+++ packages/perseus-editor/src/widgets/label-image-editor/behavior.test.tsx
