+++ packages/perseus-editor/src/widgets/label-image-editor/marker.tsx
