+++ packages/perseus-editor/src/widgets/label-image-editor/question-markers.tsx
