+++ packages/perseus-editor/src/widgets/label-image-editor/select-image.tsx
